Move all styled components to assets folder
